requirements.txt
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.14
django-filter==2.3.0
django-rest-swagger==2.2.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
html5lib==1.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==7.2.0
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
Pyphen==0.9.5
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
uritemplate==3.0.1
urllib3==1.25.10
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.2.0
